pyproject.toml
[project]
name = "skyfall"
version = "0.1.0"
description = "Real-time aerospace debris & industrial anomaly tracker"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "redis[hiredis]>=5.0.0",
    "openai>=1.14.0",
    "google-generativeai>=0.5.0",
    "telethon>=1.36.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "geopy>=2.4.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["ingestion", "processing", "output", "trajectory"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
